+++ src/tests/lib.rs
@@ -1,4 +1,4 @@
-pub fn fibbonnacci(_a: f64, _b: bool, c: f64) -> u128 {
+pub fn fibbonnacci(_a: u128, _b: bool, c: u128) -> u128 {
 
     let mut x: u128 = 0;
     let mut y: u128 = 1;
